0.0.2版本 基本完成用户管理 调整设置样式错误，样式优化 抽取公共的提示框store 公共分页total 调整为long

+++ plugins/media-fetch/m-team/src/main/java/com/yx/nas/tool/plugins/service/impl/MediaResourceFetcherImpl.java
@@ -80,7 +80,7 @@ public class MediaResourceFetcherImpl implements MediaResourceFetcher {
             PageResult<MediaResourceDto> pageResult = new PageResult<>();
             pageResult.setPageNum(reqVo.getPageNum());
             pageResult.setPageSize(reqVo.getPageSize());
-            pageResult.setTotal((Integer) parseData.get("total"));
+            pageResult.setTotal((Long) parseData.get("total"));
             pageResult.setRecords((List<MediaResourceDto>) parseData.get("data"));
 
             return pageResult;
